+++ internal/sdkprovider/userconfig/service/pg.go
@@ -42,8 +42,18 @@ func pgUserConfig() *schema.Schema {
 				Optional:    true,
 				Type:        schema.TypeInt,
 			},
+			"backup_interval_hours": {
+				Description: "Interval in hours between automatic backups. Minimum value is 3 hours. Must be a divisor of 24 (3, 4, 6, 8, 12, 24). (Applicable to ACU plans only). Example: `24`.",
+				Optional:    true,
+				Type:        schema.TypeInt,
+			},
 			"backup_minute": {
 				Description: "The minute of an hour when backup for the service is started. New backup is only started if previous backup has already completed. Example: `30`.",
+				Optional:    true,
+				Type:        schema.TypeInt,
+			},
+			"backup_retention_days": {
+				Description: "Number of days to retain automatic backups. Backups older than this value will be automatically deleted. (Applicable to ACU plans only). Example: `7`.",
 				Optional:    true,
 				Type:        schema.TypeInt,
 			},